+++ src/escape.ts
@@ -43,6 +43,27 @@ export function toUnicode(value: string): string {
 
     stringToChars(value).forEach(ch => {
         var code = ch.codePointAt(0)
+        if (code <= 0x7f) {
+            result += ch
+        } else if (code > 0xffff) {
+            let un = code - 0x10000
+            let hi = Math.floor(un / 0x400) + 0xd800
+            let lo = un % 0x400 + 0xdc00
+            result += `\\u${hi.toString(16)}\\u${lo.toString(16)}`
+        } else {
+            var codestr = ('000' + code.toString(16)).slice(-4)
+            result += `\\u${codestr}`
+        }
+    })
+
+    return result
+}
+
+export function toUnicodeAll(value: string): string {
+    let result = ''
+
+    stringToChars(value).forEach(ch => {
+        var code = ch.codePointAt(0)
         if (code > 0xffff) {
             let un = code - 0x10000
             let hi = Math.floor(un / 0x400) + 0xd800